+++ packages/ui/src/components/Stepper/__tests__/index.test.tsx
@@ -7,90 +7,90 @@ describe('Stepper', () => {
   test('renders correctly with default props', () =>
     shouldMatchEmotionSnapshot(
       <Stepper>
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly with selected prop', () =>
     shouldMatchEmotionSnapshot(
       <Stepper selected={3}>
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly with animation', () =>
     shouldMatchEmotionSnapshot(
       <Stepper animated>
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly with all selected', () =>
     shouldMatchEmotionSnapshot(
       <Stepper selected={2}>
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly with step number in row', () =>
     shouldMatchEmotionSnapshot(
       <Stepper labelPosition="right">
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly with small size', () =>
     shouldMatchEmotionSnapshot(
       <Stepper size="small">
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly without separator', () =>
     shouldMatchEmotionSnapshot(
       <Stepper separator={false}>
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly without separator with label at the right', () =>
     shouldMatchEmotionSnapshot(
       <Stepper separator={false} labelPosition="right">
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
     ))
 
   test('renders correctly with disabled steps', () =>
     shouldMatchEmotionSnapshot(
       <Stepper selected={1}>
-        <Stepper.Step index={0} title="step 1" disabled />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" disabled />
+        <Stepper.Step title="step 1" disabled />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" disabled />
       </Stepper>,
     ))
 
   test('handles clicks when interactive', () =>
     shouldMatchEmotionSnapshot(
       <Stepper selected={2} interactive>
-        <Stepper.Step index={1} title="step 1" />
-        <Stepper.Step index={2} title="step 2" />
-        <Stepper.Step index={3} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
       {
         transform: async () => {
@@ -102,9 +102,9 @@ describe('Stepper', () => {
   test('handles clicks when interactive and small', () =>
     shouldMatchEmotionSnapshot(
       <Stepper selected={2} interactive size="small">
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
       {
         transform: async () => {
@@ -116,9 +116,9 @@ describe('Stepper', () => {
   test('handles clicks when not interactive', () =>
     shouldMatchEmotionSnapshot(
       <Stepper selected={2} size="small">
-        <Stepper.Step index={0} title="step 1" />
-        <Stepper.Step index={1} title="step 2" />
-        <Stepper.Step index={2} title="step 3" />
+        <Stepper.Step title="step 1" />
+        <Stepper.Step title="step 2" />
+        <Stepper.Step title="step 3" />
       </Stepper>,
       {
         transform: async () => {
